Refactor setup script to optimize PHP extension installation
The homebrew-brewworks' setup script has been updated in order to enhance the installation process of the PHP extensions. This revision includes a modification that checks if a PHP extension is already installed before attempting to install it, which avoids unnecessary reinstallations and provides user feedback when an extension is already present.