Update phpcomplete formula with multiple PHP versions and PECL packages
This commit updates the phpcomplete formula to support multiple versions of PHP and includes the installation of XDebug and other PECL packages. These changes also include necessary dependencies, like imagemagick and libmemcached. Lastly, the formula is updated with tests to ensure the functionality of each PHP version and PECL extension.