Add section on PHP Complete in README
This commit introduces a new section in the README file, detailing how to use PHP Complete to maintain a robust PHP development environment when uninstalling BrewWorks. The section includes the benefits of using PHP Complete, as well as a link for more detailed information. This aims to help developers optimize their workflow and save time on setting up their development environment.